% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-custom-xy.r
\name{compute_com}
\alias{compute_com}
\title{Computes center of mass.
assume soma is at 0, 0, 0.
Should also have soma as input and then check this.}
\usage{
compute_com(xyz)
}
\description{
Computes center of mass.
assume soma is at 0, 0, 0.
Should also have soma as input and then check this.
}
