% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-derived.r
\name{compute_derived}
\alias{compute_derived}
\title{Requires max path distance as well. That is, it works on summarized data.
 @export}
\usage{
compute_derived(cell_vars)
}
\description{
Requires max path distance as well. That is, it works on summarized data.
 @export
}
