% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-petilla.r
\name{compute_mc_l1}
\alias{compute_mc_l1}
\title{MC upper part specific width}
\usage{
compute_mc_l1(branch_node, layer, thickness_mean, thickness_sd)
}
\description{
Considers symmetry in X (mean X), width in X, and the growth direction (mainly vertical or horizontal)
Note: for growth direction, the parent (or child) of segment may not be in the upper 165 part, but not very important
}
