% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-custom-xy.r
\name{compute_mean_direction}
\alias{compute_mean_direction}
\title{Computes mean direction
Normalizes so that each segment is given equal importance. This loses some information, as segments have different length.}
\usage{
compute_mean_direction(xyz)
}
\description{
Computes mean direction
Normalizes so that each segment is given equal importance. This loses some information, as segments have different length.
}
