% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-petilla.r
\name{compute_origin}
\alias{compute_origin}
\title{Estimate whether the axon originates above or below the soma}
\usage{
compute_origin(cell)
}
\description{
Estimate whether the axon originates above or below the soma
}
