% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-custom-xy.r
\name{compute_pca_vars}
\alias{compute_pca_vars}
\title{Compute  pca vars}
\usage{
compute_pca_vars(xyz)
}
\description{
sds, and axialized, 1st component direction
}
\details{
s1 close to s2: spherical.
s1 much higher: ellipse in 2D.
It is also possible to use SDs as variables (Cannon)
axes: 1 is x and 2 is y

radial: > 0 means radial; below 0 means going up
}
