% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-laminar.r
\name{compute_prob_translaminar}
\alias{compute_prob_translaminar}
\title{Compute translaminar}
\usage{
compute_prob_translaminar(neuron, layer, thicknesses, sds,
  assume_middle = assume_middle)
}
\description{
Compute translaminar
}
