% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives-extend.r
\name{extend_branches}
\alias{extend_branches}
\title{summarize
seems i require is terminal etc. that is for the branch to have been extended.
add that to the neurostr tools? or not?
seems it is already added.}
\usage{
extend_branches(branch)
}
\description{
summarize
seems i require is terminal etc. that is for the branch to have been extended.
add that to the neurostr tools? or not?
seems it is already added.
}
