% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{filter_dont_summarize}
\alias{filter_dont_summarize}
\title{Measures not to summarize
Contains both branch and node measures
TODO: move to neurostr!!}
\usage{
filter_dont_summarize(branch_node)
}
\description{
Measures not to summarize
Contains both branch and node measures
TODO: move to neurostr!!
}
