% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{filter_xyz}
\alias{filter_xyz}
\title{Also keep neurite type? Make it optional argument?
TODO: move to neurostr!!}
\usage{
filter_xyz(both)
}
\description{
Also keep neurite type? Make it optional argument?
TODO: move to neurostr!!
}
