% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{flag_branches}
\alias{flag_branches}
\title{TODO:  Fix this in neurostr, i.e., make them logical
#' flags to logical}
\usage{
flag_branches(branch)
}
\description{
TODO:  Fix this in neurostr, i.e., make them logical
#' flags to logical
}
