% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{get_theta_complement}
\alias{get_theta_complement}
\title{Get complementary angles for a matrix of angles
Angles muy be in degrees}
\usage{
get_theta_complement(angles, sum = 180)
}
\description{
Get complementary angles for a matrix of angles
Angles muy be in degrees
}
