% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{is_no_or_unique_neuron}
\alias{is_no_or_unique_neuron}
\title{Check a single neuron or no rows
at least move the is_unique_neuron(xyz) function with length(unique(xyz$neuron)) == 1 )}
\usage{
is_no_or_unique_neuron(primitives)
}
\description{
Check a single neuron or no rows
at least move the is_unique_neuron(xyz) function with length(unique(xyz$neuron)) == 1 )
}
