% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-laminar.r
\name{layer_to_l1}
\alias{layer_to_l1}
\title{H to pia
Sum of half L1 + above layers and optionally half its own layer}
\usage{
layer_to_l1(cell_layer, thicknesses)
}
\description{
H to pia
Sum of half L1 + above layers and optionally half its own layer
}
