% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-frontend.r
\name{quantify_branch_node}
\alias{quantify_branch_node}
\title{Quantify all for a given domain. This does not include dendritic functions, as they are a specific domain!!
Common for all domains
todo: also mean tree length could be somewhere here. fix total length?}
\usage{
quantify_branch_node(branch_node)
}
\description{
Quantify all for a given domain. This does not include dendritic functions, as they are a specific domain!!
Common for all domains
todo: also mean tree length could be somewhere here. fix total length?
}
