% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-gaba.R
\name{quantify_gaba}
\alias{quantify_gaba}
\title{A specific quantification of metrics for gabaergic interneurons}
\usage{
quantify_gaba(file, layer, thickness_mean, thickness_sd)
}
\description{
A specific quantification of metrics for gabaergic interneurons
}
