% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-frontend.r
\name{quantify_laminar}
\alias{quantify_laminar}
\title{Quantify custom vars}
\usage{
quantify_laminar(branch_node, layer, thickness_mean, thickness_sd,
  mc_axonal_pattern = TRUE)
}
\description{
Quantify custom vars
}
