% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-petilla.r
\name{quantify_petilla_neurtral}
\alias{quantify_petilla_neurtral}
\title{Petilla variables applicable to any domain that do not require laminar data.}
\usage{
quantify_petilla_neurtral(branch_node)
}
\description{
Petilla variables applicable to any domain that do not require laminar data.
}
