% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-frontend.r
\name{quantify_xyz}
\alias{quantify_xyz}
\title{Quantify xyz with node
Require node, although it really needs just neurite type and XYZ}
\usage{
quantify_xyz(node)
}
\description{
Quantify xyz with node
Require node, although it really needs just neurite type and XYZ
}
