% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-custom-dendrites.r
\name{replicate_insertion_points}
\alias{replicate_insertion_points}
\title{Replicate each point according to the length of its dendrite}
\usage{
replicate_insertion_points(lens, points)
}
\description{
Replicate each point according to the length of its dendrite
}
