% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives-extend.r
\name{summarize_measures}
\alias{summarize_measures}
\title{Summarize for a neuron}
\usage{
summarize_measures(both)
}
\value{
named numeric vector
renamed from summarize both
}
\description{
Summarize for a neuron
}
